fixed a bug when deleting not saved entity

+++ public/src/app/atendimentos/atendimento.service.ts
@@ -86,11 +86,13 @@ export class AtendimentoService {
 	}
 
 	delete(id: string) {
-		this.http.delete(`${this.url}/${id}`).subscribe(
-			res => this.dialogs.toastSuccess('Atendimento excluído com sucesso!'),
-			err => {
-				console.log(err);
-				this.dialogs.toastFail('Ocorreu um erro! Por favor, tente mais tarde.');
-		});
+		if (id) {
+			this.http.delete(`${this.url}/${id}`).subscribe(
+				res => this.dialogs.toastSuccess('Atendimento excluído com sucesso!'),
+				err => {
+					console.log(err);
+					this.dialogs.toastFail('Ocorreu um erro! Por favor, tente mais tarde.');
+			});
+		}
 	}
 }
